consumer/requirements.txt
kafka-python
# psycopg2
SQLAlchemy
psycopg2-binary